Add missing C++ Nodetemplate API methods for defining field from existing field at a node, and getting the time sequence. These are needed to interrogate time usage in the node field.
Fix return type of timesequence getNumberOfTimes method which was double but should have been an int.
Fix minor typo in timesequence C documentation.

https://tracker.physiomeproject.org/show_bug.cgi?id=3906

+++ core/source/api/zinc/timesequence.hpp
@@ -64,7 +64,7 @@ public:
 		return id;
 	}
 
-	double getNumberOfTimes()
+	int getNumberOfTimes()
 	{
 		return cmzn_timesequence_get_number_of_times(id);
 	}
